Initial commit: ripping off React TodoMVC
This is the React TodoMVC. It's an old version of React.
http://todomvc.com/examples/react/#/

I was getting an error from one of the packages in `node_modules`, but I
didn't feel like upgrading the packages or any such thing. It's on old
repo; that's not the point of it anyhow. It works as-is other than that
one line.

So I removed the offending line. And then I removed `package.json` and
renamed the `node_modules` directory to `lib` so no one would get
confused and think that there was actually some sort of package
management happening here. There's not.

This is what it is. It's a demo app. It's ready to be used as one now.